refactor(webrtc): give the channel boundary and the frame predicates an owner
The file-size ratchet fired on both webrtc files. Both grew for the same reason:
each built the channel's outbound boundary inline, which is one decision written
twice. createChannelDelivery is that decision's single owner.

pairing-gate was still over, so the three pre-accept frame predicates moved to
pairing-frames.ts — they answer what kind of frame this is, a question about the
wire vocabulary, and they were the only thing in the file not about gating. The
gate falls from 314 to 260 lines and leaves the baseline entirely; the baseline
is re-frozen in this change so the gain is locked rather than lent.
