+++ packages/agent-transport-webrtc/src/webrtc-transport.ts
@@ -1,8 +1,4 @@
-import {
-  createOutboundDelivery,
-  createWsHandler,
-  resolveAdmission,
-} from '@robota-sdk/agent-transport-protocol';
+import { createWsHandler, resolveAdmission } from '@robota-sdk/agent-transport-protocol';
 import { extractDtlsFingerprint } from '@robota-sdk/agent-remote-pairing';
 import type {
   IConfigurableTransport,
@@ -13,6 +9,7 @@ import type { RTCDataChannel, RTCPeerConnection } from 'werift';
 import type { IPairingResult } from '@robota-sdk/agent-remote-pairing';
 import type { IProtocolSession, SessionResumeBridge } from '@robota-sdk/agent-transport-protocol';
 
+import { createChannelDelivery } from './channel-delivery.js';
 import { loadWerift } from './werift-loader.js';
 import { PairingGate, type IHostReconnectConfig } from './pairing-gate.js';
 import { createTransportLifecycleError } from './transport-lifecycle-error.js';
@@ -270,13 +267,11 @@ export class WebRtcTransport implements IConfigurableTransport<IInteractiveSessi
       return;
     }
 
-    // ARCH-030: the transport is the carrier on the no-secret branch — it builds the connection's
-    // outbound boundary from its own data-channel sink and its own delivery lifecycle.
+    // ARCH-030: the transport is the carrier on the no-secret branch — its own sink, its own lifecycle.
     const { onMessage, cleanup } = createWsHandler({
       session,
-      deliver: createOutboundDelivery(
-        (serverMessage) => channel.send(JSON.stringify(serverMessage)),
-        (error, event) => this.deliveryLifecycle.handleFailure(channel, generation, error, event),
+      deliver: createChannelDelivery(channel, (error, event) =>
+        this.deliveryLifecycle.handleFailure(channel, generation, error, event),
       ),
     });
     this.cleanupHandler = cleanup;
